% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my new function.R
\name{Myfirstfunction}
\alias{Myfirstfunction}
\title{Add together two numbers}
\usage{
Myfirstfunction(x)
}
\arguments{
\item{x}{A number}
}
\value{
x, and print my sentence.
}
\description{
Add together two numbers
}
\examples{
Myfirstfunction(1,1)
}
